requirements/base.txt
Django>=5.2.0,<5.3.0
djangorestframework>=3.14.0
django-filter>=23.5
django-cors-headers>=4.3.0
django-crispy-forms>=2.0
django-extensions>=3.2.0
djangorestframework-simplejwt>=5.3.0
celery>=5.3.0
redis>=5.0.0
gunicorn>=20.1.0
psycopg[binary]>=3.2.0
whitenoise>=6.6.0
Pillow>=10.0.0
python-dotenv>=1.0.0
django-environ>=0.11.0
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
reportlab>=4.0.0
PyYAML>=6.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
python-magic>=0.4.27
django-auditlog>=3.0.0
django-axes>=6.1.0
django-ratelimit>=4.1.0
django-mptt>=0.16.0
